adding test for geoNear aggregation

+++ src/de/caluga/morphium/query/geospatial/LineString.java
@@ -3,7 +3,7 @@ package de.caluga.morphium.query.geospatial;
 import java.util.ArrayList;
 import java.util.List;
 
-public class LineString extends Geo<List<int[]>> {
+public class LineString extends Geo<List<double[]>> {
     public LineString() {
         super(GeoType.LINESTRING);
         setCoordinates(new ArrayList<>());
