+++ src/de/caluga/morphium/query/geospatial/MultiLineString.java
@@ -3,7 +3,7 @@ package de.caluga.morphium.query.geospatial;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MultiLineString extends Geo<List<List<int[]>>> {
+public class MultiLineString extends Geo<List<List<double[]>>> {
     public MultiLineString() {
         super(GeoType.MULITLINESTRING);
         setCoordinates(new ArrayList<>());
