+++ src/de/caluga/morphium/query/geospatial/MultiPolygon.java
@@ -3,7 +3,7 @@ package de.caluga.morphium.query.geospatial;
 import java.util.ArrayList;
 import java.util.List;
 
-public class MultiPolygon extends Geo<List<List<List<int[]>>>> {
+public class MultiPolygon extends Geo<List<List<List<double[]>>>> {
     public MultiPolygon() {
         super(GeoType.MULTIPOLYGON);
         setCoordinates(new ArrayList<>());
