+++ src/de/caluga/morphium/query/geospatial/Polygon.java
@@ -7,7 +7,7 @@ import java.util.List;
 // list of polygon rings
 // list of coordinates
 // list of points
-public class Polygon extends Geo<List<List<int[]>>> {
+public class Polygon extends Geo<List<List<double[]>>> {
     public Polygon() {
         super(GeoType.POLYGON);
         setCoordinates(new ArrayList<>());
